+++ src/composables/use-search.ts
@@ -23,13 +23,17 @@ export default function useSearch({ ctx }: Options) {
   ) => {
     apiState.fetching = true
 
+    const Cookie = process.client ? require('js-cookie') : undefined
+    const cookieLang = process.client ? (Cookie.get('i18n_redirected') ? Cookie.get('i18n_redirected') : null) : null
+
     const { data } = await ctx.root.$axios.get(`${process.env.NUXT_ENV_WORDPRESS_API_URL}/wp-json/wp/v2/search`, {
       params: {
         per_page: 10,
         search: searchContent,
         page: infiniteLoadingPage + 1,
         type: type,
-        subtype: subtype
+        subtype: subtype,
+        lang: cookieLang
       }
     })
     globalState.searchResults = data
